make

+++ libft/ft_calloc.c
@@ -6,13 +6,14 @@
 /*   By: mdor <mdor@student.s19.be>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 18:47:38 by mdor              #+#    #+#             */
-/*   Updated: 2022/12/08 12:42:04 by mdor             ###   ########.fr       */
+/*   Updated: 2022/12/11 12:19:08 by mdor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+//#include "libft.h"
+#include <stdlib.h>
 
-void *ft_calloc(size_t elemCount, size_t elemSize)
+void *ft_calloc(int elemCount, int elemSize)
 {
     void *a;
