+++ libft/ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: mdor <mdor@student.s19.be>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 18:51:27 by mdor              #+#    #+#             */
-/*   Updated: 2022/12/07 14:44:52 by mdor             ###   ########.fr       */
+/*   Updated: 2022/12/11 12:20:12 by mdor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void    *ft_memcpy(void *destination, const void *source, size_t n)
+void    *ft_memcpy(void *destination, const void *source, int n)
 {
     char    *dst;
     const char *src;
-    size_t  i;
+    int  i;
 
     i = 0;
     if (n == 0 || destination == source)
